requirements.txt
absl-py==2.3.1
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.14.3
black==24.8.0
bleach==6.1.0
builtin-interfaces==1.2.2
cachetools==5.5.2
ccimport==0.3.7
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
composition-interfaces==1.2.2
contourpy==1.1.1
crcmod==1.7
cryptography==46.0.3
cumm-cu111==0.2.9
cv-bridge==3.2.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
demo-nodes-py==0.20.5
descartes==1.1.0
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.14.0
fire==0.7.1
flake8==7.1.2
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.0
geometry-msgs==4.9.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
image-geometry==3.2.1
imageio @ file:///croot/imageio_1707247282708/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
interactive-markers==2.3.2
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.7
lark==1.3.1
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lifecycle-msgs==1.2.2
llvmlite==0.31.0
logging-demo==0.20.5
lyft-dataset-sdk==0.0.8
map-msgs==2.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
message-filters==4.3.8
mistune==3.1.4
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv-full==1.4.0
mmdet==2.14.0
# Editable Git install with no remote (mmdet3d==0.17.1)
-e /mnt/EC0A5E060A5DCE68/Ubuntu22_04/UniData/mmdetection3d
mmsegmentation==0.14.1
model-index==0.1.11
mypy_extensions==1.1.0
narwhals==1.42.1
nav-msgs==4.9.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
ninja==1.13.0
notebook==7.3.3
notebook_shim==0.2.4
numba==0.48.0
numpy @ file:///work/mkl/numpy_and_numpy_base_1682950891414/work
nuscenes-devkit==1.1.9
oauthlib==3.3.1
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
osrf-pycommon==2.1.6
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pccm==0.3.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.5.0
pluggy==1.5.0
plyfile==1.0.3
portalocker==3.0.0
prettytable==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.23
pycryptodome==3.23.0
pyflakes==3.2.0
Pygments==2.19.2
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-qt-binding==1.1.2
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.3
pyzmq==27.1.0
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
referencing==0.35.1
requests==2.32.4
requests-oauthlib==2.0.0
resource-retriever==3.1.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rmw-dds-common==1.6.0
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.20.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.3
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spconv-cu111==2.1.25
sros2==0.10.7
stack-data==0.6.3
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
tabulate==0.9.0
teleop-twist-keyboard==2.4.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
tf2-geometry-msgs==0.25.15
tf2-kdl==0.25.15
tf2-msgs==0.25.15
tf2-py==0.25.15
tf2-ros-py==0.25.15
tf2-sensor-msgs==0.25.15
tf2-tools==0.25.15
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.22
tinycss2==1.2.1
tomli==2.3.0
topic-monitor==0.20.5
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
trajectory-msgs==4.9.0
trimesh==2.35.39
turtlesim==1.4.2
typing_extensions==4.13.2
unique-identifier-msgs==2.2.1
uri-template==1.3.0
urllib3==1.26.20
visualization-msgs==4.9.0
wcwidth==0.2.14
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.15
yapf==0.43.0
zipp==3.20.2
